refactor(directory): move preprocess to directory service
- remove rawdirectory search from urlrepository
- shift preprocess task to directorysearchservice
